requirements.txt
# Core FastAPI and Web Framework
fastapi>=0.119.1
uvicorn>=0.30.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI and Language Processing
langchain==1.0.1
langchain-openai==1.0.1
langchain-community==0.4
langchain-core==1.0.0
langchain-chroma==1.0.0
openai>=1.109.1
langfuse==2.51.2

# Vector Database and Search
chromadb>=1.0.20

# HTTP and Web Scraping
requests>=2.32.5
httpx>=0.27.0
beautifulsoup4==4.12.2

# Data Processing and Validation
pydantic>=2.7.4
pandas==2.3.3

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf4llm==0.0.5
python-docx==1.1.0
docx2txt==0.8

# Browser Automation for SharePoint
selenium==4.36.0
webdriver-manager==4.0.2

# Excel Processing
openpyxl==3.1.2
xlrd==2.0.1

# Text Processing
markdown==3.5.1

# Database
motor==3.3.2
pymongo==4.5.0

# JSON-based memory storage (no external dependencies needed)

# Enhanced Document Processing
unstructured[all-docs]==0.11.8
pytesseract==0.3.10
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'

# Machine Learning and NLP
scikit-learn==1.4.0
spacy==3.7.2
tiktoken==0.5.2

# Additional utilities
Pillow==10.2.0